Fix error that was causing Scala compiler crash.
Oddly, the crash was in the bytecode emitter phase, and named
PorcERuntime.scala as the problem, but the problem was just syntax
errors in InvokeCallRecordNode.java.

The project builds now.

+++ PorcE/src/orc/run/porce/CallBase.java
@@ -40,7 +40,7 @@ public abstract class CallBase extends Expression {
 		if (e instanceof HasPorcNode && ((HasPorcNode) e).porcNode().isDefined()) {
 			PorcAST ast = ((HasPorcNode) e).porcNode().get();
 			if (ast instanceof ASTWithIndex && ((ASTWithIndex) ast).optionalIndex().isDefined()) {
-				return ((ASTWithIndex) ast).optionalIndex().get();
+				return (Integer) ((ASTWithIndex) ast).optionalIndex().get();
 			}
 		}
 		Node p = e.getParent();
